g4f/.v1/pyproject.toml
[tool.poetry]
name = "gpt4free"
version = "0.1.0"
description = ""
authors = []
license = "GPL-3.0"
readme = "README.md"
packages = [{ include = "gpt4free" }]
exclude = ["**/*.txt"]

[tool.poetry.dependencies]
python = "^3.10"

colorama = "^0.4.6"
curl-cffi = "^0.5.5"
fake-useragent = "^1.1.3"
Levenshtein = "*"
mailgw_temporary_email = "*"
names = "^0.3.0"
pycryptodome = "*"
pydantic = "^1.10.7"
pypasser = "^0.0.5"
pymailtm = "*"
random-password-generator = "*"
requests = "^2.29.0"
retrying = "*"
selenium = "^4.9.0"
streamlit-chat = { git = "https://github.com/AI-Yash/st-chat", rev = "ffb0583" }
streamlit = "^1.21.0"
tls-client = "^0.2"
twocaptcha = "^0.0.1"
websocket-client = "^1.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
